requirements.txt
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.4
defusedxml==0.7.1
et_xmlfile==2.0.0
fonttools==4.61.1
fpdf2==2.8.5
numpy==2.4.1
openpyxl==3.1.5
pandas==3.0.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.0
pycparser==3.0
pypdfium2==5.3.0
python-dateutil==2.9.0.post0
six==1.17.0
